requirements.txt
fastapi
uvicorn
openai
python-dotenv
sentence-transformers
faiss-cpu
pypdf
numpy